requirements/prod.txt
Flask
flask-sqlalchemy
psycopg2-binary
flask-migrate
gunicorn
alembic 
